Difference in inkscape files.
I added an verification in generate.py file, because some files
had the size on picture at attribute viewBox and some others in
width and height.

I had a problem with some pictures but all is now fine.

This  link show a picture with the differences of both files.

http://dl.dropbox.com/u/32744064/Pantallazo%20del%202012-05-21%2001%3A38%3A42.png

Sorry for my english, i'm new with this languaje.

	modified:   generate.py